ImplicitAD

Documentation for ImplicitAD.

ImplicitAD.drdy_forward

ImplicitAD.implicit_function

ImplicitAD.implicit_linear_function

ImplicitAD.jvp_forward

ImplicitAD.jvp_provide

ImplicitAD.linear_solve

ImplicitAD.pack_dual

ImplicitAD.unpack_dual

ImplicitAD.vjp_provide

ImplicitAD.vjp_reverse

ImplicitAD.drdy_forward — Method

compute Aij = ∂ri/∂y_j

ImplicitAD.implicit_function — Method

implicit_function(solve, residual, x; jvp=jvp_forward, vjp=vjp_reverse, drdx=nothing, drdy=drdy_forward, lsolve=linear_solve)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments

solve::function: y = solve(x). Solve implicit function (where residual!(r, x, y) = 0, see below)

residual!::function: residual!(r, x, y). Set residual r, given input x and state y.

x::vector{float}: evaluation point.

jvp::function: jvp(residual, x, y, v). Compute Jacobian vector product b = B*v where Bij = ∂ri/∂x_j. r = residual(x, y) (note explicit form. Default leverages dual numbers in forward mode AD where Jacobian is not explicitly constructed.

vjp::function: vjp(residual, x, y, v). Compute vector Jacobian product c = B^T v = (v^T B)^T where Bij = ∂ri/∂x_j and return c Default is reverse mode AD where Jacobian is not explicitly constructed. Computes gradient of a scalar function.

drdx::function: drdx(residual, x, y). Provide (or compute yourself): ∂ri/∂xj. Not a required function. Default is nothing. If provided, then jvp and jvp are ignored and explicitly multiplied against this provided Jacobian.

drdy::function: drdy(residual, x, y). Provide (or compute yourself): ∂ri/∂yj. Default is forward mode AD.

lsolve::function: lsolve(A, b). Linear solve A x = b (where A is computed in drdy and b is computed in jvp, or it solves A^T x = c where c is computed in vjp). Default is backslash operator.

ImplicitAD.implicit_linear_function — Method

implicit_linear_function(A, b; lsolve=linear_solve, fact=factorize)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff). This version is for linear equations Ay = b

Arguments

A::matrix, b::vector: $y = A^{-1} b$

lsolve::function: lsolve(A, b). Linear solve A y = b

ImplicitAD.jvp_forward — Method

Compute Jacobian vector product b = B*v where Bij = ∂ri/∂x_j This takes in the dual directly since it is already formed that way.

ImplicitAD.jvp_provide — Method

internal convenience method for when user can provide drdx

ImplicitAD.linear_solve — Method

Linear solve A x = b (where A is computed in drdy and b is computed in jvp).

ImplicitAD.pack_dual — Method

Create a ForwardDiff Dual with value yv, derivatives dy, and Dual type T

ImplicitAD.unpack_dual — Method

unpack ForwardDiff Dual return value and derivative.

ImplicitAD.vjp_provide — Method

internal convenience method for when user can provide drdx

ImplicitAD.vjp_reverse — Method

Compute vector Jacobian product c = B^T v = (v^T B)^T where Bij = ∂ri/∂x_j and return c