ImplicitAD

Documentation for ImplicitAD.

ImplicitAD.drdy_forward

ImplicitAD.implicit_function

ImplicitAD.jvp_forward

ImplicitAD.jvp_provide

ImplicitAD.linear_solve

ImplicitAD.pack_dual

ImplicitAD.transpose_linear_solve

ImplicitAD.unpack_dual

ImplicitAD.vjp_provide

ImplicitAD.vjp_reverse

ImplicitAD.drdy_forward — Method

compute Aij = ∂ri/∂y_j

ImplicitAD.implicit_function — Method

implicit_function(solve, residual, x; jvp=jvp_forward, vjp=vjp_reverse, drdx=nothing, drdy=drdy_forward, lsolve=linear_solve, tlsolve=transpose_linear_solve)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments

solve::function: y = solve(x). Solve implicit function (where residual(x, y) = 0) given input x and return states y(x).

residual::function: r = residual(x, y). Return residual given input x and state y.

x::vector{float}: evaluation point.

jvp::function: jvp(residual, x, y, v). Compute Jacobian vector product b = B*v where Bij = ∂ri/∂x_j. Default is forward mode AD where jacobian is not explicitly constructed.

vjp::function: vjp(residual, x, y, v). Compute vector Jacobian product c = B^T v = (v^T B)^T where Bij = ∂ri/∂x_j and return c Default is reverse mode AD where Jacobian is not explicitly constructed.

drdx::function: drdx(residual, x, y). Provide (or compute yourself): ∂ri/∂xj. Not a required function. Default is nothing. If provided, then jvp and jvp are ignored and explicitly multiplied against this provided Jacobian.

drdy::function: drdy(residual, x, y). Provide (or compute yourself): ∂ri/∂yj. Default is forward mode AD.

lsolve::function: lsolve(A, b). Linear solve A x = b (where A is computed in drdy and b is computed in jvp). Default is backslash operator.

tlsolve::function: tlsolve(A, b). Transpose linear solve A^T x = c (where A is computed in drdy and c is computed in vjp). Default is forwardslash operator.

ImplicitAD.jvp_forward — Method

Compute Jacobian vector product b = B*v where Bij = ∂ri/∂x_j

ImplicitAD.jvp_provide — Method

internal convenience method for when user can provide drdx

ImplicitAD.linear_solve — Method

Linear solve A x = b (where A is computed in drdy and b is computed in jvp).

ImplicitAD.pack_dual — Method

Create a ForwardDiff Dual with value yv, derivatives dy, and Dual type T

ImplicitAD.transpose_linear_solve — Method

transpose linear solve A^T x = c (where A is computed in drdy and c is computed in vjp)

ImplicitAD.unpack_dual — Method

unpack ForwardDiff Dual return value and derivative.

ImplicitAD.vjp_provide — Method

internal convenience method for when user can provide drdx

ImplicitAD.vjp_reverse — Method

Compute vector Jacobian product c = B^T v = (v^T B)^T where Bij = ∂ri/∂x_j and return c