ImplicitAD

Documentation for ImplicitAD.

ImplicitAD.computeA

ImplicitAD.implicit_function

ImplicitAD.jvp

ImplicitAD.pack_dual

ImplicitAD.unpack_dual

ImplicitAD.vjp

ImplicitAD.computeA — Method

compute Aij = ∂ri/∂y_j

ImplicitAD.implicit_function — Method

implicit_function(solve, residual, x, drdy=nothing, drdx=nothing)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments

solve::function: y = solve(x). Solve implicit function (where residual(x, y) = 0) given input x and return states y(x).

residual::function: r = residual(x, y). Return residual given input x and state y.

x::vector{float}: evaluation point.

drdy::function: ∂r/∂y = drdy(x, y). Defaults to nothing. Set to a function if you can provide the partials directly.

drdx::function: ∂r/∂x = drdx(x, y). Defaults to nothing. Set to a function if you can provide the partials directly.

ImplicitAD.jvp — Method

Compute Jacobian vector product b = B*v where Bij = ∂ri/∂x_j

ImplicitAD.pack_dual — Method

Create a ForwardDiff Dual with value yv, derivatives dy, and Dual type T

ImplicitAD.unpack_dual — Method

unpack ForwardDiff Dual return value and derivative.

ImplicitAD.vjp — Method

Compute vector Jacobian product c = B^T v = (v^T B)^T where Bij = ∂ri/∂x_j and return c