Reference

ImplicitAD.implicit — Function

implicit(solve, residual, x, p=(); drdy=drdy_forward, lsolve=linear_solve)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments

solve::function: y = solve(x, p). Solve implicit function returning state variable y, for input variables x, and fixed parameters p.

residual::function: Either r = residual(y, x, p) or in-place residual!(r, y, x, p). Set residual r (scalar or vector), given state y (scalar or vector), variables x (vector) and fixed parameters p (tuple).

x::vector{float}: evaluation point.

p::tuple: fixed parameters. default is empty tuple.

drdy::function: drdy(residual, y, x, p). Provide (or compute yourself): ∂ri/∂yj. Default is forward mode AD.

lsolve::function: lsolve(A, b). Linear solve A x = b (where A is computed in drdy and b is computed in jvp, or it solves A^T x = c where c is computed in vjp). Default is backslash operator.

ImplicitAD.implicit_linear — Function

implicit_linear(A, b; lsolve=linear_solve, fact=factorize)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff). This version is for linear equations Ay = b

Arguments

A::matrix, b::vector: components of linear system $A y = b$

lsolve::function: lsolve(A, b). Function to solve the linear system, default is backslash operator.

Af::factorization: An optional factorization of A, useful to override default factorize, or if multiple linear solves will be performed with same A matrix.

ImplicitAD.apply_factorization — Function

apply_factorization(A, factfun)

Apply a matrix factorization to the primal portion of a dual number. Avoids user from needing to add ForwardDiff as a dependency. Afactorization = factfun(A)

ImplicitAD.provide_rule — Function

provide_rule(func, x, p=(); mode="ffd", jacobian=nothing, jvp=nothing, vjp=nothing)

Provide partials rather than rely on AD. For cases where AD is not available or to provide your own rule, or to use mixed mode, etc.

Arguments

func::function, x::vector{float}, p::tuple: function of form: y = func(x, p), where x are variables and p are fixed parameters.

mode::string:

"ffd": forward finite difference

"cfd": central finite difference

"cs": complex step

"jacobian": provide your own jacobian (see jacobian function)

"vp": provide jacobian vector product and vector jacobian product (see jvp and vjp)

jacobian::function: only used if mode="jacobian". J = jacobian(x, p) provide Jij = dyi / dx_j

jvp::function: only used if mode="vp" and in forward mode. ydot = jvp(x, p, v) provide Jacobian vector product J*v

vjp::function: only used if mode="vp" and in revese mode. xbar = vjp(x, p, v) provide vector Jacobian product v'*J