Reference

ImplicitAD.implicit — Function

implicit(solve, residual, x, p=(); drdy=drdy_forward, lsolve=linear_solve)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments

solve::function: y = solve(x, p). Solve implicit function returning state variable y, for input variables x, and fixed parameters p.

residual::function: Either r = residual(y, x, p) or in-place residual!(r, y, x, p). Set residual r (scalar or vector), given state y (scalar or vector), variables x (vector) and fixed parameters p (tuple).

x::vector{float}: evaluation point.

p::tuple: fixed parameters. default is empty tuple.

drdy::function: drdy(residual, y, x, p). Provide (or compute yourself): ∂ri/∂yj. Default is forward mode AD.

lsolve::function: lsolve(A, b). Linear solve A x = b (where A is computed in drdy and b is computed in jvp, or it solves A^T x = c where c is computed in vjp). Default is backslash operator.

ImplicitAD.implicit_linear — Function

implicit_linear(A, b; lsolve=linear_solve, Af=factorize)

Make implicit function AD compatible (specifically with ForwardDiff and ReverseDiff). This version is for linear equations Ay = b

Arguments

A::matrix, b::vector: components of linear system $A y = b$

lsolve::function: lsolve(A, b). Function to solve the linear system, default is backslash operator.

Af::factorization: An optional factorization of A, useful to override default factorize, or if multiple linear solves will be performed with same A matrix.

ImplicitAD.apply_factorization — Function

apply_factorization(A, factfun)

Apply a matrix factorization to the primal portion of a dual number. Avoids user from needing to add ForwardDiff as a dependency. Afactorization = factfun(A)

ImplicitAD.implicit_eigval — Function

implicit_eigval(A, B, eigsolve)

Make eigenvalue problems AD compatible with ForwardDiff and ReverseDiff

Arguments

A::matrix, B::matrix: generlized eigenvalue problem. A v = λ B v (B is identity for standard eigenvalue problem)

eigsolve::function: λ, V, U = eigsolve(A, B). Function to solve the eigenvalue problem. λ is a vector containing eigenvalues. V is a matrix whose columns are the corresponding eigenvectors (i.e., λ[i] corresponds to V[:, i]). U is a matrix whose columns contain the left eigenvectors (u^H A = λ u^H B) The left eigenvectors must be in the same order as the right ones (i.e., U' * B * V must be diagonal). U can be provided with any normalization as we normalize internally s.t. U' * B * V = I If A and B are symmetric/Hermitian then U = V.

Returns

λ::vector: eigenvalues and their derivatives. (Currently only eigenvalue derivatives are provided. not eigenvectors)

ImplicitAD.explicit_unsteady — Function

explicit_unsteady(solve, residual, x, p=(); drdy=drdy_forward, lsolve=linear_solve)

Make an explicit time-marching analysis AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments:

solve::function: function y, t = solve(x, p). Perform a time marching analysis, and return the matrix y = [y[1] y[2] y[3] ... y[N]] wherey[i]is the state vector at time stepi(a rows is a state, a columms is a timesteps) and vector t = [t[1], t[2], t[3] ... t[N]] where t[i] is the corresponding time, for input variables x, and fixed variables p.

perform_step::function: Either y[i] = perform_step(y[i-1], t[i], t[i-1], x, p) or in-place perform_step!(y[i], y[i-1], t[i], t[i-1], x, p). Set the next set of state variables y[i] (scalar or vector), given the previous state y[i-1] (scalar or vector), current time t[i], previous time t[i-1], variables x (vector), and fixed parameters p (tuple).

x: evaluation point

p: fixed parameters, default is empty tuple.

ImplicitAD.implicit_unsteady — Function

implicit_unsteady(solve, residual, x, p=(); drdy=drdy_forward, lsolve=linear_solve)

Make a implicit time-marching analysis AD compatible (specifically with ForwardDiff and ReverseDiff).

Arguments:

solve::function: function y, t = solve(x, p). Perform a time marching analysis, and return the matrix y = [y[1] y[2] y[3] ... y[N]] wherey[i]is the state vector at time stepi(a rows is a state, a columms is a timesteps) and vector t = [t[1], t[2], t[3] ... t[N]] where t[i] is the corresponding time, for input variables x and fixed variables p.

residual::function: Either r[i] = residual(y[i], y[i-1], t[i], t[i-1], x, p) or in-place residual!(r[i], y[i], y[i-1], t[i], t[i-1], x, p). Set current residual r[i] (scalar or vector), given current state y[i] (scalar or vector), previous state y[i-1] (scalar or vector), current and preivous time variables t[i] and t[i-1], x (vector), and fixed parameters p (tuple).

x: evaluation point

p: fixed parameters, default is empty tuple.

drdy: drdy(residual, y, x, p). Provide (or compute yourself): ∂ri/∂yj. Default is forward mode AD.

lsolve::function: lsolve(A, b). Linear solve A x = b (where A is computed in drdy and b is computed in jvp, or it solves A^T x = c where c is computed in vjp). Default is backslash operator.

ImplicitAD.provide_rule — Function

provide_rule(func, x, p=(); mode="ffd", jacobian=nothing, jvp=nothing, vjp=nothing)

Provide partials rather than rely on AD. For cases where AD is not available or to provide your own rule, or to use mixed mode, etc.

Arguments

func::function, x::vector{float}, p::tuple: function of form: y = func(x, p), where x are variables and p are fixed parameters.

mode::string:

"ffd": forward finite difference

"cfd": central finite difference

"cs": complex step

"jacobian": provide your own jacobian (see jacobian function)

"vp": provide jacobian vector product and vector jacobian product (see jvp and vjp)

jacobian::function: only used if mode="jacobian". J = jacobian(x, p) provide Jij = dyi / dx_j

jvp::function: only used if mode="vp" and in forward mode. ydot = jvp(x, p, v) provide Jacobian vector product J*v

vjp::function: only used if mode="vp" and in revese mode. xbar = vjp(x, p, v) provide vector Jacobian product v'*J